Wait until we've got the ready msg from the server
The code didn't wait until it got the "220 ready" message from
the server, and just sent the command once the connection was
established. This resulted in the callback being called to early
with the status code 220 when talking to a real server (the tests
just sent 220 message with the rest of the data). That's now fixed and
the test now behave more like a real SVDRP server.